poddley_transcriber_and_crawler/requirements.txt
clint==0.5.1
feedparser==6.0.10
langcodes==3.3.0
openai-whisper==20230314
requests==2.28.2
safe_cast==0.3.4
whisper==1.1.10
prisma
pandas
stable-ts